compilation warnings fixes

+++ library/tulip-gui/include/tulip/cxx/TulipItemEditorCreators.cxx
@@ -60,7 +60,7 @@ QWidget *NumberEditorCreator<T>::createWidget(QWidget *parent) const {
   } else if (typeid(T).name() == typeid(tlp::UnsignedIntegerType).name()) {
     dsb->setRange(0, UINT_MAX);
   } else if (typeid(T).name() == typeid(tlp::LongType).name()) {
-    dsb->setRange(-LONG_MAX, LONG_MAX);
+    dsb->setRange(static_cast<double>(-LONG_MAX), static_cast<double>(LONG_MAX));
   } else if (typeid(T).name() == typeid(tlp::FloatType).name()) {
     dsb->setRange(-FLT_MAX, FLT_MAX);
   } else {
